Support multiple tests for hp_core module
A test is composed of an RISC-V assembly file with the instructions to
execute and a .yaml file with the same name containing the expected
values in the registers of interest after execution. Each test will be
compiled into a HEX file that is loaded by the imem module and will
produce its own wave file.